+++ bcbio/ngsalign/star.py
@@ -62,6 +62,7 @@ def align(fastq_file, pair_file, ref_file, names, align_dir, data):
         cmd += " --readFilesCommand zcat " if is_gzipped(fastq_file) else ""
         cmd += _read_group_option(names)
         fusion_mode = utils.get_in(data, ("config", "algorithm", "fusion_mode"), False)
+        fusion_mode = fusion_mode or dd.get_fusion_caller(data)
         if fusion_mode:
             cmd += (" --chimSegmentMin 12 --chimJunctionOverhangMin 12 "
                     "--chimScoreDropMax 30 --chimSegmentReadGapMax 5 "